requirements.txt
Flask==2.3.2
flask-restx==1.1.0
flask_sqlalchemy==3.0.5
psycopg2-binary==2.9.6
Flask-Migrate==4.0.4
PyJWT==2.7.0